# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pickleball Club Ladder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List invitations for current player
      responses:
        "200":
          description: Invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationsResponse"
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Create a team invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationBody"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInvitation"
  /invitations/{id}/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept a team invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation accepted, team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithPlayers"
  /invitations/{id}/decline:
    post:
      operationId: declineInvitation
      tags: [invitations]
      summary: Decline a team invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # LADDER
  /ladder:
    get:
      operationId: getLadder
      tags: [ladder]
      summary: Get full ladder standings for active season
      parameters:
        - name: season_id
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ladder standings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LadderResponse"
...
  /ladder/{id}/position:
    patch:
      operationId: updateLadderPosition
      tags: [ladder]
      summary: Manually update a team's ladder position (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePositionBody"
      responses:
        "200":
          description: Position updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LadderStandingWithTeam"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth
...
    Player:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [player, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
...
    Season:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - startDate
        - endDate
        - isActive
        - createdAt
...
    Team:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        player1Id:
          type: string
        player2Id:
          type: string
        teamName:
          type: string
        status:
          type: string
          enum: [pending, active, inactive]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - player1Id
        - player2Id
        - teamName
        - status
        - createdAt
    TeamWithPlayers:
      allOf:
        - $ref: "#/components/schemas/Team"
        - type: object
          properties:
            player1:
              $ref: "#/components/schemas/Player"
            player2:
              $ref: "#/components/schemas/Player"
            standing:
              $ref: "#/components/schemas/LadderStanding"
            season:
              $ref: "#/components/schemas/Season"
...
    TeamInvitation:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        inviterId:
          type: string
        inviteeId:
          type: string
        teamName:
          type: string
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
        inviter:
          $ref: "#/components/schemas/Player"
        invitee:
          $ref: "#/components/schemas/Player"
        season:
          $ref: "#/components/schemas/Season"
      required:
        - id
        - seasonId
        - inviterId
        - inviteeId
        - teamName
        - status
        - createdAt
    CreateInvitationBody:
      type: object
      properties:
        inviteeId:
          type: string
        teamName:
          type: string
      required:
        - inviteeId
        - teamName
    InvitationsResponse:
      type: object
      properties:
        sent:
          type: array
          items:
            $ref: "#/components/schemas/TeamInvitation"
        received:
          type: array
          items:
            $ref: "#/components/schemas/TeamInvitation"
      required:
        - sent
        - received

    # Ladder
    LadderStanding:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        teamId:
          type: string
        position:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        lastMatchDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - teamId
        - position
        - wins
        - losses
        - createdAt
    LadderStandingWithTeam:
      allOf:
        - $ref: "#/components/schemas/LadderStanding"
        - type: object
          properties:
            team:
              $ref: "#/components/schemas/TeamWithPlayers"
    LadderResponse:
      type: object
      properties:
        season:
          $ref: "#/components/schemas/Season"
        standings:
          type: array
          items:
            $ref: "#/components/schemas/LadderStandingWithTeam"
        total:
          type: integer
      required:
        - standings
        - total
...
    UpdatePositionBody:
      type: object
      properties:
        position:
          type: integer
      required:
        - position
...